+++ main.py
@@ -15,7 +15,8 @@
 #                                                                             #
 # Developpers :                                                               #
 # - Noah JACOB                                                                #
-# - Oliver MINEAU                                                             #
+# - Oliver MINEAU  
+# - Eirik Lu                                                          #
 #                                                                             #
 ###############################################################################
 
@@ -80,7 +81,7 @@ from dev.devices.sim_spectrometer import MySimSpectrometer
 
 
 # Define the application version
-version = "V3.0.2"
+version = "V3.0.3"
 
 # Set the size of the application's main window (width, height)
 size = (1200,700)
@@ -93,13 +94,14 @@ visible_notif_time = 3
 backup_directory = r"C:\Users\el-ei\Documents\Masterproject\Backup data"
 
 # Create an instance of a microscope setup
+# Remove 'Sim' from the class names to connect to real devices 
 gm_microscope = MyMicroscope("Maria Goeppert Mayer",
     MySimCamera("Camera", "28939", enable=True),
     MySimSpectrometer("Spectrometer-VIS", "QEP06226", enable=True, dark_correction=True, integration_time=100),
     MySimSpectrometer("Spectrometer-NIR", "NQ51B1981", integration_time=50),
     MySimRotationMount("White-Light-Plzr", "COM6", "QEP06226", enable=False),
     MyRotationMount("Laser-Plzr", "COM7", "QEP06226", enable=False),
-    MyShutter("White light - shutter", "26006167", enable=True, model="KST201"),
+    MyShutter("White light - shutter", "26006167", enable=True, model="KST201", stage="FW103M"),
     MyShutter("Laser 750nm - shutter", "68801094", enable=False, model="KSC101"),
     MyShutter("Laser 550nm - shutter", "68800970", enable=False, model="KSC101"),
     MyFilter("Filter 12", "xxxx"),
@@ -113,7 +115,7 @@ gm_microscope = MyMicroscope("Maria Goeppert Mayer",
 lm_microscope = MyMicroscope("Lise Meitner",
     MySpectrometer("Spectrometer", "xxxx", enable=True),
     MyCamera("Camera", "11499", enable=True),
-    MyShutter("White light - shutter", "26006167", enable=True, model="KST201"),
+    MyShutter("White light - shutter", "26006167", enable=True, model="KST201", stage="FW103M"),
     MyShutter("Laser 750nm - shutter", "68801094", enable=False, model="KSC101"),
     MyFilter("Filter 12", "xxxx"),
     MyStage("Stage", "xxxx"))